Refactored(step-1)
