+++ src/components/Main/Post/Post.css
