--- src/components/Plug/Plug.css
@@ -1,41 +1,0 @@
-.plug__container {
-  margin: 10px auto;
-  position: relative;
-  width: 90%;
-}
-
-.plug__image {
-  display: block;
-  object-fit: cover;
-  width: 100%;
-  height: 50px;
-}
-
-.plug__overlay {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  height: 100%;
-  width: 100%;
-  opacity: 0;
-  transition: .5s ease;
-  background-color: #008CBA;
-}
-
-.plug__container:hover .plug__overlay {
-  opacity: 1;
-}
-
-.plug__text {
-  color: white;
-  font-size: 20px;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-  text-align: center;
-}
